Revert refactor of parallelization_wrappers.py
